requirements.txt
testcontainers==4.7.1
pytest~=8.2.2
requests==2.32.3
Flask==3.0.3
Flask-SQLAlchemy==3.0.4
scapy==2.5.0
kafka-python-ng==2.2.2
confluent-kafka==2.5.0
six==1.16.0
sqlalchemy==2.0.32
python-dotenv==1.0.0
pandas==2.2.2
